+++ stac_fastapi/sfeos_helpers/stac_fastapi/sfeos_helpers/database/__init__.py
@@ -56,6 +56,7 @@ from .mapping import get_queryables_mapping_shared
 from .query import (
     apply_collections_bbox_filter_shared,
     apply_collections_datetime_filter_shared,
+    apply_collections_free_text_filter_shared,
     apply_free_text_filter_shared,
     apply_intersects_filter_shared,
     populate_sort_shared,
@@ -87,6 +88,7 @@ __all__ = [
     "apply_intersects_filter_shared",
     "apply_collections_bbox_filter_shared",
     "apply_collections_datetime_filter_shared",
+    "apply_collections_free_text_filter_shared",
     "populate_sort_shared",
     # Mapping operations
     "get_queryables_mapping_shared",